refactor: 스토리 리포지토리 추상화
