+++ src/modules/extensions/essay/story/story.module.ts
@@ -1,12 +1,12 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { StoryService } from './story.service';
-import { StoryRepository } from './story.repository';
+import { StoryService } from './core/story.service';
+import { StoryRepository } from './infrastructure/story.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Story } from '../../../../entities/story.entity';
 import { ToolModule } from '../../../utils/tool/tool.module';
 import { UserModule } from '../../../base/user/user.module';
 import { EssayModule } from '../../../base/essay/essay.module';
-import { StoryController } from './story.controller';
+import { StoryController } from './api/story.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthModule } from '../../../base/auth/auth.module';
 
@@ -20,7 +20,7 @@ import { AuthModule } from '../../../base/auth/auth.module';
     forwardRef(() => AuthModule),
   ],
   controllers: [StoryController],
-  providers: [StoryService, StoryRepository],
-  exports: [StoryService, StoryRepository],
+  providers: [StoryService, { provide: 'IStoryRepository', useClass: StoryRepository }],
+  exports: [StoryService, { provide: 'IStoryRepository', useClass: StoryRepository }],
 })
 export class StoryModule {}
