+++ src/modules/extensions/essay/story/tests/story.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { StoryController } from '../story.controller';
-import { StoryService } from '../story.service';
+import { StoryController } from '../api/story.controller';
+import { StoryService } from '../core/story.service';
 import { AuthGuard } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule } from '@nestjs/config';
@@ -8,7 +8,7 @@ import { Request as ExpressRequest } from 'express';
 import { CreateStoryReqDto } from '../dto/repuest/createStoryReq.dto';
 import { UpdateStoryReqDto } from '../dto/repuest/updateStoryReq.dto';
 
-jest.mock('../story.service');
+jest.mock('../core/story.service');
 
 describe('StoryController', () => {
   let controller: StoryController;
